melhorando padrão e fluxo do código
