--- views/agenda/excluir.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Excluir Aluno</title>
-</head>
-<body>
-    <form id="alunoFormExcluir">
-        <label for="nome">ID: </label>
-        <input type="number" name="id">
-        <input type="submit" value="Excluir">
-    </form>
-    <a href="../index.html">Voltar</a>
-    <script src="../scriptAgenda.js"></script>
-</body>
-</html>
