+++ views/agenda/incluir.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
 
 <head>
     <meta charset="UTF-8">
@@ -9,7 +9,7 @@
 
 <body>
     <form id="agenda-form-incluir">
-        <input type="text" name="aluno_id" value="" hidden readonly disabled required>
+        <input type="hidden" name="aluno_id" value="" required>
 
         <label for="nome">Nome do Aluno </label>
         <input type="text" name="nome" readonly disabled required>
@@ -29,7 +29,7 @@
         <input type="submit">
     </form>
     <a href="../index.html">Voltar</a>
-    <script src="../assets/scripts/agenda/script.js" type="module"></script>
+    <script src="../assets/scripts/agenda/incluir.js" type="module"></script>
 </body>
 
 </html>
