+++ views/alunos/atualizarAluno.html
@@ -1,12 +1,12 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Incluir Aluno</title>
 </head>
 <body>
-    <form id="alunoFormAtualizar">
+    <form id="aluno-atualizar-form">
         <label for="id">ID: </label>
         <input type="number" name="id">
         <label for="nome">Nome: </label>
