+++ views/alunos/incluir.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
 
 <head>
     <meta charset="UTF-8">
@@ -8,7 +8,7 @@
 </head>
 
 <body>
-    <form id="alunoFormIncluir">
+    <form id="aluno-form-incluir">
         <label for="nome">Nome: </label>
         <input type="text" name="nome" required>
         
@@ -24,7 +24,7 @@
         <input type="submit" value="Incluir">
     </form>
     <a href="../index.html">Voltar</a>
-    <script src="../assets/scripts/alunos/incluir.js"></script>
+    <script src="../assets/scripts/alunos/incluir.js" type="module"></script>
 </body>
 
 </html>
