+++ views/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
 
 <head>
     <meta charset="UTF-8">
@@ -11,18 +11,14 @@
     <h1>Agenda Alunos</h1>
 
     <div>
-        <a href="./alunos/incluirAluno.html">Incluir Aluno</a>
-        <a href="./alunos/listarAlunos.html">Listar Alunos</a>
+        <a href="./alunos/incluir.html">Incluir Aluno</a>
+        <a href="./alunos/listar.html">Listar Alunos</a>
         <a href="./alunos/atualizarAluno.html">Atualizar Aluno</a>
         <a href="./alunos/excluirAluno.html">Excluir Aluno</a>
     </div>
 
     <div>
-        <a href="./agenda_alunos/incluir.html">Incluir Agenda</a>
-        <a href="./agenda_alunos/listar.html">Listar Agenda</a>
-        <a href="./agenda_alunos/excluir.html">Atualizar Agenda</a>
-        <a href="./agenda_alunos/excluir.html">Excluir Agenda</a>
-
+        <a href="./agenda/listar_excluir.html">Agendas</a>
     </div>
 </body>
